Solved challenge 1
Prompt: Reverse an array. A classic code interview question. 
Difficulty: easy
Languages: no-restrictions
Constraints: The function should take in any array and return that array in reverse. If your language has a native .reverse() function cannot use it because this is a test of complexity and not how well you can use a function!